Changes for 0.9.0.1-maemo0

+++ build/make-maemo-debuild.sh
@@ -10,20 +10,26 @@ if [ ! -e $SOURCE ]; then
 	echo "Source path does not exist; Exiting!"
 	exit
 fi
+# Check if dest exists
+if [ -e $PKG ]; then
+	echo "Package path exists; Exiting!"
+	echo "To delete, run rm -r $PKG"
+	exit
+fi
 # Create destination dir if necessary
-mkdir -p $PKG
+#mkdir -p $PKG
 mkdir -p $PKGTMP
 
 cp -a $PKGROOT/* $PKGTMP/
 mkdir -p $PKGTMP/src/opt/agtl-maemo/
 #sed -e "s/___VERSION___/$VERSION/" $PKGROOT/setup.py > $PKGTMP/setup.py
 # Copy python sources 
-cp $SOURCE/utils.py $SOURCE/astral.py $SOURCE/connection.py $SOURCE/gpsreader.py $SOURCE/cachedownloader.py $SOURCE/coordfinder.py $SOURCE/geo.py $SOURCE/gui.py $SOURCE/cli.py $SOURCE/core.py $SOURCE/geocaching.py $SOURCE/provider.py $SOURCE/colorer.py $SOURCE/downloader.py $SOURCE/geonames.py $SOURCE/hildongui.py $SOURCE/simplegui.py $SOURCE/hildon_plugins.py $SOURCE/gtkmap.py $SOURCE/abstractmap.py $SOURCE/openstreetmap.py $SOURCE/portrait.py $SOURCE/threadpool.py $PKGTMP/src/opt/agtl-maemo/
+rsync -av --delete --exclude='*.pyc' $SOURCE/utils.py $SOURCE/astral.py $SOURCE/connection.py $SOURCE/gpsreader.py $SOURCE/cachedownloader.py $SOURCE/coordfinder.py $SOURCE/geo.py $SOURCE/gui.py $SOURCE/cli.py $SOURCE/core.py $SOURCE/geocaching.py $SOURCE/provider.py $SOURCE/colorer.py $SOURCE/downloader.py $SOURCE/geonames.py $SOURCE/hildongui.py $SOURCE/simplegui.py $SOURCE/hildon_plugins.py $SOURCE/gtkmap.py $SOURCE/abstractmap.py $SOURCE/openstreetmap.py $SOURCE/portrait.py $SOURCE/threadpool.py $PKGTMP/src/opt/agtl-maemo/
 # Copy additional resources
 cp -r $SOURCE/data $PKGTMP/src/opt/agtl-maemo/
 cp -r $SOURCE/actors $PKGTMP/src/opt/agtl-maemo/
 cd $PKGTMP/
 #PSA_ROOT=$PYSIDE_ROOT $PYSIDE_ASSISTANT --project $PYSIDE_PROJECT_NAME build-deb
-debuild
+debuild -aarmel
 #cd -
 #cp $PKGTMP/deb_dist/${PYSIDE_PROJECT_NAME}_${VERSION}-1_all.deb $PKG
